Trying to get Mario Land working correctly: * Reduced copying in GPU with array segments * Reduced complexity in MMU * Reduced complexity and stuttering caused by async interrupts. * GPU is now reactive. * Removed loads of boilerplate in API by switching to protobuf. * Switched to FluentAssertions and Bogus instead of Shouldly and custom thing.

+++ Retro.Net/Z80/Memory/CacheAwareZ80Mmu.cs
@@ -1,8 +1,10 @@
 ﻿using Retro.Net.Memory;
 using Retro.Net.Memory.Dma;
+using Retro.Net.Memory.Interfaces;
 using Retro.Net.Timing;
-using Retro.Net.Z80.Cache;
+using Retro.Net.Z80.Cache.Interfaces;
 using Retro.Net.Z80.Config;
+using Retro.Net.Z80.Core.Interfaces;
 using Retro.Net.Z80.Peripherals;
 
 namespace Retro.Net.Z80.Memory
@@ -16,19 +18,20 @@ namespace Retro.Net.Z80.Memory
         private readonly IInstructionBlockCache _instructionBlockCache;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="CacheAwareZ80Mmu"/> class.
+        /// Initializes a new instance of the <see cref="CacheAwareZ80Mmu" /> class.
         /// </summary>
         /// <param name="peripheralManager">The peripheral manager.</param>
         /// <param name="platformConfig">The platform configuration.</param>
         /// <param name="memoryBankController">The memory bank controller.</param>
         /// <param name="dmaController">The dma controller.</param>
         /// <param name="instructionTimer">The instruction timer.</param>
+        /// <param name="instructionBlockCache">The instruction block cache.</param>
         public CacheAwareZ80Mmu(IPeripheralManager peripheralManager,
-            IPlatformConfig platformConfig,
-            IMemoryBankController memoryBankController,
-            IDmaController dmaController,
-            IInstructionTimer instructionTimer,
-            IInstructionBlockCache instructionBlockCache)
+                                IPlatformConfig platformConfig,
+                                IMemoryBankController memoryBankController,
+                                IDmaController dmaController,
+                                IInstructionTimer instructionTimer,
+                                IInstructionBlockCache instructionBlockCache)
             : base(peripheralManager, platformConfig, memoryBankController, dmaController, instructionTimer)
         {
             _instructionBlockCache = instructionBlockCache;
@@ -39,6 +42,6 @@ namespace Retro.Net.Z80.Memory
         /// </summary>
         /// <param name="address"></param>
         /// <param name="length"></param>
-        protected override void OnAddressWrite(ushort address, ushort length) => _instructionBlockCache.InvalidateCache(address, length);
+        protected override void OnAddressWrite(ushort address, int length) => _instructionBlockCache.InvalidateCache(address, length);
     }
 }
